solana_zk_token_sdk::instruction

Module transfer

Structs

FeeEncryption

FeeParameters

TransferAmountCiphertext

TransferData

The instruction data that is needed for the ProofInstruction::VerifyTransfer instruction.

TransferProofContext

The context data needed to verify a transfer proof.

TransferPubkeys

The ElGamal public keys needed for a transfer

TransferWithFeeData

The instruction data that is needed for the ProofInstruction::TransferWithFee instruction.

TransferWithFeeProofContext

The context data needed to verify a transfer-with-fee proof.

TransferWithFeePubkeys

The ElGamal public keys needed for a transfer with fee

Enums

Role

Functions

combine_lo_hi_commitments

combine_lo_hi_openings

combine_lo_hi_u64

split_u64

Takes in a 64-bit number amount and a bit length bit_length. It returns: