Struct solana_zk_token_sdk::instruction::transfer_with_fee::TransferWithFeeData

#[repr(C)]
pub struct TransferWithFeeData {
    pub context: TransferWithFeeProofContext,
    pub proof: TransferWithFeeProof,
}

Fields
context: TransferWithFeeProofContext

The context data for the transfer with fee proof

proof: TransferWithFeeProof
Implementations

impl TransferWithFeeData

pub fn new( transfer_amount: u64, (spendable_balance, old_source_ciphertext): (u64, &ElGamalCiphertext), source_keypair: &ElGamalKeypair, (destination_pubkey, auditor_pubkey): (&ElGamalPubkey, &ElGamalPubkey), fee_parameters: FeeParameters, withdraw_withheld_authority_pubkey: &ElGamalPubkey ) -> Result<Self, ProofError>

fn ciphertext_lo(&self, role: Role) -> Result<ElGamalCiphertext, ProofError>

Extracts the lo ciphertexts associated with a transfer-with-fee data

fn ciphertext_hi(&self, role: Role) -> Result<ElGamalCiphertext, ProofError>

Extracts the lo ciphertexts associated with a transfer-with-fee data

fn fee_ciphertext_lo(&self, role: Role) -> Result<ElGamalCiphertext, ProofError>

Extracts the lo fee ciphertexts associated with a transfer_with_fee data

fn fee_ciphertext_hi(&self, role: Role) -> Result<ElGamalCiphertext, ProofError>

Extracts the hi fee ciphertexts associated with a transfer_with_fee data

pub fn decrypt_amount( &self, role: Role, sk: &ElGamalSecretKey ) -> Result<u64, ProofError>

Decrypts transfer amount from transfer-with-fee data

pub fn decrypt_fee_amount( &self, role: Role, sk: &ElGamalSecretKey ) -> Result<u64, ProofError>

Decrypts transfer amount from transfer-with-fee data