solana_zk_token_sdk::zk_token_elgamal

Module pod

Structs

AeCiphertext

The AeCiphertext type as a Pod.

BatchedGroupedCiphertext2HandlesValidityProof

The BatchedGroupedCiphertext2HandlesValidityProof type as a Pod.

CiphertextCiphertextEqualityProof

The CiphertextCiphertextEqualityProof type as a Pod.

CiphertextCommitmentEqualityProof

The CiphertextCommitmentEqualityProof type as a Pod.

CompressedRistretto

DecryptHandle

The DecryptHandle type as a Pod.

ElGamalCiphertext

The ElGamalCiphertext type as a Pod.

ElGamalPubkey

The ElGamalPubkey type as a Pod.

FeeEncryption

FeeParameters

FeeSigmaProof

The FeeSigmaProof type as a Pod.

GroupedCiphertext2HandlesValidityProof

The GroupedCiphertext2HandlesValidityProof type as a Pod.

GroupedElGamalCiphertext2Handles

The GroupedElGamalCiphertext type with two decryption handles as a Pod

GroupedElGamalCiphertext3Handles

The GroupedElGamalCiphertext type with three decryption handles as a Pod

PedersenCommitment

The PedersenCommitment type as a Pod.

PodProofType

PodU16

PodU64

PubkeyValidityProof

The PubkeyValidityProof type as a Pod.

RangeProofU64

The RangeProof type as a Pod restricted to proofs on 64-bit numbers.

RangeProofU128

The RangeProof type as a Pod restricted to proofs on 128-bit numbers.

RangeProofU256

The RangeProof type as a Pod restricted to proofs on 256-bit numbers.

TransferAmountCiphertext

ZeroBalanceProof

The ZeroBalanceProof type as a Pod.

Traits

Pod

Marker trait for “plain old data”.

Zeroable

Trait for types that can be safely created with zeroed.

Derive Macros

Pod

Derive the Pod trait for a struct

Zeroable

Derive the Zeroable trait for a type.